chore: remove rusoto2
