+++ src/sinks/util/mod.rs
@@ -5,10 +5,8 @@ pub mod http;
 pub mod http2;
 pub mod retries;
 pub mod retries2;
-#[cfg(feature = "rusoto_core")]
-pub mod rusoto;
 #[cfg(feature = "rusoto_core44")]
-pub mod rusoto2;
+pub mod rusoto;
 pub mod service;
 pub mod service2;
 pub mod sink;
